+++ src/Libraries/RevitNodes/Properties/Resources.Designer.cs
@@ -304,6 +304,15 @@ namespace Revit.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to Current Document is not workshared.
+        /// </summary>
+        internal static string DocumentNotWorkshared {
+            get {
+                return ResourceManager.GetString("DocumentNotWorkshared", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Dynamo Data.
         /// </summary>
         internal static string Dynamo_AVF_Data_Name {
@@ -732,6 +741,14 @@ namespace Revit.Properties {
         internal static string NoParentElement {
             get {
                 return ResourceManager.GetString("NoParentElement", resourceCulture);
+            }
+        }
+
+        ///   Looks up a localized string similar to No Elements to purge in the current document..
+        /// </summary>
+        internal static string NoElementsToPurge {
+            get {
+                return ResourceManager.GetString("NoElementsToPurge", resourceCulture);
             }
         }