+++ test/Libraries/RevitIntegrationTests/DocumentTests.cs
@@ -1,4 +1,6 @@
-﻿using System.IO;
+﻿using System.Collections.Generic;
+using System.IO;
+using System.Linq;
 
 using NUnit.Framework;
 
@@ -11,7 +13,7 @@ using RTF.Framework;
 namespace RevitSystemTests
 {
     [TestFixture]
-    class DocumentTests : RevitSystemTestBase
+    public class DocumentTests : RevitSystemTestBase
     {
         [Test]
         [TestModel(@"./empty.rfa")]
@@ -73,6 +75,98 @@ namespace RevitSystemTests
         public void WhenActiveDocumentResetIsRequiredVisualizationsAreCleared()
         {
             Assert.Inconclusive("Cannot test. API required for allowing closing all docs.");
+        }
+
+        [Test]
+        [TestModel(@".\element.rvt")]
+        public void CanSaveFamilyInCurrentDocument()
+        {
+            // Arange
+            string samplePath = Path.Combine(workingDirectory, @".\Document\canSaveFamiliesInCurrentDocument.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+
+            string expectedSavedFamilyFileName = "Rectangular Column.rfa";
+            int expectedSavedFileCount = 1;
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+
+            string savedFamilyDirectoryPath = GetPreviewValue("0af5cc47acab47369c479f3a4b198306").ToString();
+
+            List<string> files = Directory.GetFiles(savedFamilyDirectoryPath).Select(x => Path.GetFileName(x)).ToList();
+            int resultSavedFileCount = files.Count;
+            string savedFamilyName = files.FirstOrDefault();
+            var fileExistInFolder = File.Exists(Path.GetFullPath(Path.Combine(savedFamilyDirectoryPath, expectedSavedFamilyFileName)));
+
+            // Assert
+            Assert.AreEqual(expectedSavedFileCount, resultSavedFileCount);
+            Assert.AreEqual(savedFamilyName, expectedSavedFamilyFileName);
+            Assert.IsTrue(fileExistInFolder);
+
+            // Clean up
+            Directory.Delete(savedFamilyDirectoryPath, true);
+        }
+
+        [Test]
+        [TestModel(@".\Document\LocalModel\Project1_LocalFile.rvt")]
+        public void CanGetWorksharingModelPath()
+        {
+            // Arrange
+            string samplePath = Path.Combine(workingDirectory, @".\Document\canGetWorksharingModelPath.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+            string expectedWorksharingFilePath = @"DynamoRevit\test\System\Document\CentralModel";
+            bool expectedIsCloudPathResult = false;
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+            var resultWorksharingPath = GetPreviewValue("c699cecf62434bcba194378b819af876") as string;
+            var resultIsCloudPath = GetPreviewValue("1b62b04935b84f58a31bf45efe48955d");
+
+            // Assert
+            Assert.IsTrue(resultWorksharingPath.Contains(expectedWorksharingFilePath));
+            Assert.AreEqual(expectedIsCloudPathResult, resultIsCloudPath);
+        }
+
+        [Test]
+        [TestModel(@".\Document\BIM360\4481adfb-0f03-4e58-9f49-8bd37dde9e0e.rvt")]
+        public void CanGetWorksharingModelPathOnCloudModel()
+        {
+            // Arrange
+            string samplePath = Path.Combine(workingDirectory, @".\Document\canGetWorksharingModelPath.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+            string expectedWorksharingFilePath = @"BIM 360://Node test/BIM360_model.rvt";
+            bool expectedIsCloudPathResult = true;
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+
+            var resultWorksharingPath = GetPreviewValue("c699cecf62434bcba194378b819af876") as string;
+            var resultIsCloudPath = GetPreviewValue("1b62b04935b84f58a31bf45efe48955d");
+
+            // Assert
+            Assert.IsTrue(resultWorksharingPath.Contains(expectedWorksharingFilePath));
+            Assert.AreEqual(expectedIsCloudPathResult, resultIsCloudPath);
+        }
+
+        [Test]
+        [TestModel(@".\Document\DocumentPurgeUnusedTest.rvt")]
+        public void CanPurgeUnusedElementsFromDocument()
+        {
+            // Arange
+            string samplePath = Path.Combine(workingDirectory, @".\Document\canPurgeUnusedElementsFromDocument.dyn");
+            string testPath = Path.GetFullPath(samplePath);
+            var expectedPurgedElementIds = new List<int>() { 217063, 221347, 216753, 416, 208080, 210695 };
+
+            // Act
+            ViewModel.OpenCommand.Execute(testPath);
+            RunCurrentModel();
+            var purgeUnusedOutput = GetPreviewCollection("7997eedbf6bd4822b5ca8b8cf0819de2");
+
+            // Assert
+            CollectionAssert.AreEqual(purgeUnusedOutput, expectedPurgedElementIds);
         }
     }
 }