+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.2.0
+* Fix issue - Joined geometry does not convert correctly
+* Document nodes - Add 3 node in the Application.Document category
+* Update Dynamo Core to 2.5.0.7432
+
+## 0.1.13
+* update DynamoCore to 2.5.0.7186
+
+## 0.1.12
+* Remove transient elements in Revit doc when Dynamo shutdown
+* Fix test failure of JsonRestoresBinding
+
 ## 0.1.11
 * Element nodes second batch - add 5 new nodes for DynamoRevit